requirements_pip_freeze.txt
certifi==2021.10.8
charset-normalizer==2.0.6
configparser==5.0.2
idna==3.2
numpy==1.19.5
pandas==1.1.5
pathlib2==2.3.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
python-dotenv==0.19.1
